+++ arenas.html
@@ -8,6 +8,8 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>MK - JSMarathon - Arenas</title>
     <link rel="stylesheet" href="style.css">
+    <!--<audio src='assets/sounds/arenas_2.mp3' autoplay='true' loop='true'></audio>-->
+    <script defer src="main.js" type="module"></script>
 </head>
 
 <body>
@@ -59,8 +61,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <audio src='assets/sounds/arenas_2.mp3' autoplay='true' loop='true'></audio>
-    <script src="main.js" type="module"></script>
 </body>
 
 </html>
