+++ index.css
@@ -62,6 +62,24 @@ body {
     height: 100%;
 }
 
+.enemy__player {
+    position: absolute;
+    width: 210px;
+    height: 367px;
+    z-index: 3;
+    bottom: 44px;
+    right: -27%;
+    transform: translate(-205%, 0px);
+    background: blue;
+    opacity: 0.2;
+}
+
+.enemy__player img {
+    max-width: 100%;
+    max-height: 100%;
+    height: 100%;
+}
+
 .parent {
     display: grid;
     grid-template-columns: repeat(7, 1fr);
