+++ index.html
@@ -8,6 +8,8 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Mortal Kombat</title>
     <link rel="stylesheet" href="index.css">
+    <!--<audio src='assets/sounds/choose.mp3' autoplay='true' loop='true'></audio>-->
+    <script defer src="index.js" type='module'></script>
 </head>
 
 <body>
@@ -16,10 +18,9 @@
             SELECT YOUR FIGHTERS
         </div>
         <div class="player"></div>
+        <div class="enemy__player"></div>
         <div class="parent"></div>
     </div>
-    <audio src='assets/sounds/choose.mp3' autoplay='true' loop='true'></audio>
-    <script src="index.js" type='module'></script>
 </body>
 
 </html>
